New Wizard Power mechanics; spell changes - Store a WizardPower property in the WizardPlayer class, instead of using the WizardBar as a pseudo-variable. - Change WizardPower to use a true 0 to 1000 point scale (instead of the 0 to 1 point scale of the WizardBar) - Change MagicChest to allow ItemStacks of more than 1 item to enact MagicPatterns - Add command for setting WizardPower - Expand WizardPlayer to allow multiple spell timers - Change patterns to use GLOWSTONE_DUST instead of REDSTONE - Change tools to use 1 Wizard Power per use instead of 1 per block removed - Add WizardFood patterns and handlers - Add HungerForPower spell - Change Incinerate to a Projectile spell

+++ me/mistergone/AWizardDidIt/MagicSpell.java
@@ -7,9 +7,10 @@ import java.util.ArrayList;
 
 public class MagicSpell {
     protected String spellName;
-    protected double cost;
+    protected int cost;
     protected ArrayList<String> reagents;
     protected SpellFunction spellFunction;
+
 
     public String getSpellName() {
         return this.spellName;
